Moves flushPromiseQueue into DevTools
