+++ src/Apps/Artwork/Components/__tests__/PricingContextModal.test.tsx
@@ -3,7 +3,7 @@ import { mockTracking } from "Artsy/Analytics"
 import { mount } from "enzyme"
 import React from "react"
 import { PricingContextModal } from "../PricingContextModal"
-import { flushPromiseQueue } from "Utils/flushPromiseQueue"
+import { flushPromiseQueue } from "DevTools"
 
 jest.unmock("react-relay")
 jest.unmock("react-tracking")
