+++ src/DevTools/index.ts
@@ -4,3 +4,4 @@ export * from "./MockRelayRenderer"
 export * from "./MockRouter"
 export * from "./renderRelayTree"
 export * from "./renderUntil"
+export * from "./flushPromiseQueue"
